refactor(ui): move What's New version modules into versions folder
